+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.0.7 (WIP)
+
+- Simple coloring when a program is running
+- Fix odd value rom length crash when viewing instruction mode
+
 ## 0.0.6
 
-- Fix bug in `jp` mnemonic when using `org`
+- Fix bug in `jp` mnemonic when using `.org`
 
 ## 0.0.5